+++ lua/oil/init.lua
@@ -129,7 +129,7 @@ M.set_sort = function(sort)
 end
 
 ---Change how oil determines if the file is hidden
----@param is_hidden_file fun(filename: string, bufnr: nil|integer): boolean Return true if the file/dir should be hidden
+---@param is_hidden_file fun(filename: string, bufnr: integer): boolean Return true if the file/dir should be hidden
 M.set_is_hidden_file = function(is_hidden_file)
   require("oil.view").set_is_hidden_file(is_hidden_file)
 end
@@ -140,12 +140,14 @@ M.toggle_hidden = function()
 end
 
 ---Get the current directory
+---@param bufnr? integer
 ---@return nil|string
-M.get_current_dir = function()
+M.get_current_dir = function(bufnr)
   local config = require("oil.config")
   local fs = require("oil.fs")
   local util = require("oil.util")
-  local scheme, path = util.parse_url(vim.api.nvim_buf_get_name(0))
+  local buf_name = vim.api.nvim_buf_get_name(bufnr or 0)
+  local scheme, path = util.parse_url(buf_name)
   if config.adapters[scheme] == "files" then
     assert(path)
     return fs.posix_to_os_path(path)
@@ -286,29 +288,22 @@ M.open_float = function(dir)
     })
   )
 
-  -- Update the window title when we switch buffers
-  if vim.fn.has("nvim-0.9") == 1 and config.float.border ~= "none" then
-    local function get_title()
-      local src_buf = vim.api.nvim_win_get_buf(winid)
-      local title = vim.api.nvim_buf_get_name(src_buf)
-      local scheme, path = util.parse_url(title)
-      if config.adapters[scheme] == "files" then
-        assert(path)
-        local fs = require("oil.fs")
-        title = vim.fn.fnamemodify(fs.posix_to_os_path(path), ":~")
-      end
-      return title
-    end
-    table.insert(
-      autocmds,
-      vim.api.nvim_create_autocmd("BufWinEnter", {
-        desc = "Update oil floating window title when buffer changes",
-        pattern = "*",
-        callback = function(params)
-          local winbuf = params.buf
-          if not vim.api.nvim_win_is_valid(winid) or vim.api.nvim_win_get_buf(winid) ~= winbuf then
-            return
-          end
+  table.insert(
+    autocmds,
+    vim.api.nvim_create_autocmd("BufWinEnter", {
+      desc = "Reset local oil window options when buffer changes",
+      pattern = "*",
+      callback = function(params)
+        local winbuf = params.buf
+        if not vim.api.nvim_win_is_valid(winid) or vim.api.nvim_win_get_buf(winid) ~= winbuf then
+          return
+        end
+        for k, v in pairs(config.float.win_options) do
+          vim.api.nvim_set_option_value(k, v, { scope = "local", win = winid })
+        end
+
+        -- Update the floating window title
+        if vim.fn.has("nvim-0.9") == 1 and config.float.border ~= "none" then
           local cur_win_opts = vim.api.nvim_win_get_config(winid)
           vim.api.nvim_win_set_config(winid, {
             relative = "editor",
@@ -316,12 +311,12 @@ M.open_float = function(dir)
             col = cur_win_opts.col,
             width = cur_win_opts.width,
             height = cur_win_opts.height,
-            title = get_title(),
+            title = util.get_title(winid),
           })
-        end,
-      })
-    )
-  end
+        end
+      end,
+    })
+  )
 
   vim.cmd.edit({ args = { util.escape_filename(parent_url) }, mods = { keepalt = true } })
   -- :edit will set buflisted = true, but we may not want that
@@ -415,11 +410,14 @@ M.close = function()
   vim.api.nvim_buf_delete(oilbuf, { force = true })
 end
 
+---@class oil.OpenPreviewOpts
+---@field vertical? boolean Open the buffer in a vertical split
+---@field horizontal? boolean Open the buffer in a horizontal split
+---@field split? "aboveleft"|"belowright"|"topleft"|"botright" Split modifier
+
 ---Preview the entry under the cursor in a split
----@param opts nil|table
----    vertical boolean Open the buffer in a vertical split
----    horizontal boolean Open the buffer in a horizontal split
----    split "aboveleft"|"belowright"|"topleft"|"botright" Split modifier
+---@param opts? oil.OpenPreviewOpts
+---@param callback? fun(err: nil|string) Called once the preview window has been opened
 M.open_preview = function(opts, callback)
   opts = opts or {}
   local config = require("oil.config")