+++ lua/oil/util.lua
@@ -8,6 +8,8 @@ local FIELD_NAME = constants.FIELD_NAME
 local FIELD_TYPE = constants.FIELD_TYPE
 local FIELD_META = constants.FIELD_META
 
+---@alias oil.IconProvider fun(type: string, name: string, conf: table?): (icon: string, hl: string)
+
 ---@param url string
 ---@return nil|string
 ---@return nil|string
@@ -358,6 +360,26 @@ M.is_floating_win = function(winid)
   return vim.api.nvim_win_get_config(winid or 0).relative ~= ""
 end
 
+---Recalculate the window title for the current buffer
+---@param winid nil|integer
+---@return string
+M.get_title = function(winid)
+  if config.float.get_win_title ~= nil then
+    return config.float.get_win_title(winid or 0)
+  end
+
+  local src_buf = vim.api.nvim_win_get_buf(winid or 0)
+  local title = vim.api.nvim_buf_get_name(src_buf)
+  local scheme, path = M.parse_url(title)
+
+  if config.adapters[scheme] == "files" then
+    assert(path)
+    local fs = require("oil.fs")
+    title = vim.fn.fnamemodify(fs.posix_to_os_path(path), ":~")
+  end
+  return title
+end
+
 local winid_map = {}
 M.add_title_to_win = function(winid, opts)
   opts = opts or {}
@@ -365,21 +387,10 @@ M.add_title_to_win = function(winid, opts)
   if not vim.api.nvim_win_is_valid(winid) then
     return
   end
-  local function get_title()
-    local src_buf = vim.api.nvim_win_get_buf(winid)
-    local title = vim.api.nvim_buf_get_name(src_buf)
-    local scheme, path = M.parse_url(title)
-    if config.adapters[scheme] == "files" then
-      assert(path)
-      local fs = require("oil.fs")
-      title = vim.fn.fnamemodify(fs.posix_to_os_path(path), ":~")
-    end
-    return title
-  end
   -- HACK to force the parent window to position itself
   -- See https://github.com/neovim/neovim/issues/13403
   vim.cmd.redraw()
-  local title = get_title()
+  local title = M.get_title(winid)
   local width = math.min(vim.api.nvim_win_get_width(winid) - 4, 2 + vim.api.nvim_strwidth(title))
   local title_winid = winid_map[winid]
   local bufnr
@@ -427,7 +438,7 @@ M.add_title_to_win = function(winid, opts)
         if vim.api.nvim_win_get_buf(winid) ~= winbuf then
           return
         end
-        local new_title = get_title()
+        local new_title = M.get_title(winid)
         local new_width =
           math.min(vim.api.nvim_win_get_width(winid) - 4, 2 + vim.api.nvim_strwidth(new_title))
         vim.api.nvim_buf_set_lines(bufnr, 0, -1, true, { " " .. new_title .. " " })
@@ -665,6 +676,20 @@ M.get_preview_win = function()
   end
 end
 
+---@return fun() restore Function that restores the cursor
+M.hide_cursor = function()
+  vim.api.nvim_set_hl(0, "OilPreviewCursor", { nocombine = true, blend = 100 })
+  local original_guicursor = vim.go.guicursor
+  vim.go.guicursor = "a:OilPreviewCursor/OilPreviewCursor"
+
+  return function()
+    -- HACK: see https://github.com/neovim/neovim/issues/21018
+    vim.go.guicursor = "a:"
+    vim.cmd.redrawstatus()
+    vim.go.guicursor = original_guicursor
+  end
+end
+
 ---@param bufnr integer
 ---@param preferred_win nil|integer
 ---@return nil|integer
@@ -841,6 +866,33 @@ M.get_edit_path = function(bufnr, entry, callback)
     adapter.get_entry_path(url, entry, callback)
   else
     adapter.normalize_url(url, callback)
+  end
+end
+
+--- Check for an icon provider and return a common icon provider API
+---@return (oil.IconProvider)?
+M.get_icon_provider = function()
+  -- prefer mini.icons
+  local _, mini_icons = pcall(require, "mini.icons")
+  ---@diagnostic disable-next-line: undefined-field
+  if _G.MiniIcons then -- `_G.MiniIcons` is a better check to see if the module is setup
+    return function(type, name)
+      return mini_icons.get(type == "directory" and "directory" or "file", name)
+    end
+  end
+
+  -- fallback to `nvim-web-devicons`
+  local has_devicons, devicons = pcall(require, "nvim-web-devicons")
+  if has_devicons then
+    return function(type, name, conf)
+      if type == "directory" then
+        return conf and conf.directory or "", "OilDirIcon"
+      else
+        local icon, hl = devicons.get_icon(name)
+        icon = icon or (conf and conf.default_file or "")
+        return icon, hl
+      end
+    end
   end
 end